refactor: make currency as an enum class
