+++ src/main/kotlin/christmas/domain/money/Currency.kt
@@ -1,0 +1,6 @@
+package christmas.domain.money
+
+enum class Currency {
+    KRW,
+    ;
+}
