+++ src/main/kotlin/christmas/domain/money/Money.kt
@@ -1,8 +1,8 @@
-package christmas.domain
+package christmas.domain.money
 
 import java.math.BigDecimal
 
 data class Money(
     val amount: BigDecimal,
-    val currency: String
+    val currency: Currency
 )
