+++ src/main/kotlin/christmas/domain/product/Product.kt
@@ -1,6 +1,6 @@
 package christmas.domain.product
 
-import christmas.domain.Money
+import christmas.domain.money.Money
 
 data class Product(
     val productType: ProductType,
